% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipart-upload-api.R
\name{AbortMultipartUpload}
\alias{AbortMultipartUpload}
\title{AbortMultipartUpload}
\usage{
AbortMultipartUpload(bucketname, key, uploadId)
}
\arguments{
\item{uploadId}{Multipart upload event ID. InitiateMultipartUpload will return this in the response.}
}
\description{
AbortMultipartUpload
}
\examples{
AbortMultipartUpload('ross-test', 'multi-test.txt', 'F7376AAB033344C1922A588D15CE56A2')
}
