% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{aclBucket}
\alias{aclBucket}
\title{aclBucket}
\usage{
aclBucket(bucketname, acl)
}
\arguments{
\item{acl}{Authority control of bucket. choose from public-read-write, public-read, private.}
}
\description{
aclBucket
}
\examples{
aclBucket('ross-test')
aclBucket('ross-test', 'public-read')
}
