% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{aclMultipleObjects}
\alias{aclMultipleObjects}
\title{aclMultipleObjects}
\usage{
aclMultipleObjects(bucketname, prefix, acl, recursive = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
aclMultipleObjects
}
\examples{
PutObject('ross-test', 'test1.txt')
PutObject('ross-test', 'test2.txt')
aclMultipleObjects('ross-test', 'test1.txt', 'public-read', .progressbar=F)
aclObject('ross-test', 'test1.txt')
aclMultipleObjects('ross-test', 'test', 'public-read-write', recursive = T, .progressbar=F)
aclObject('ross-test', 'test2.txt')
}
