% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{aclObject}
\alias{aclObject}
\title{aclObject}
\usage{
aclObject(bucketname, key, acl)
}
\arguments{
\item{acl}{}
}
\description{
aclObject
}
\examples{
aclObject('ross-test', 'test.txt')
aclObject('ross-test', 'test.txt', 'public-read')
}
