\name{big.eight}
\alias{big.eight}
\docType{data}
\title{Example dataset for Ross Ordering}
\description{Dataset contains 8 stimuli listing the names of universities that made up the Big Nine conference (without Iowa).}
\usage{data(big.eight)}
\format{
  A data frame with 8 observations on the following variable.
  \describe{
    \item{\code{University}}{a factor with levels \code{Indiana University} \code{Northwestern University} \code{University of Chicago} \code{University of Illinois} \code{University of Iowa} \code{University of Michigan} \code{University of Minnesota} \code{University of Wisconsin}}
  }
}

\source{\url{http://en.wikipedia.org/wiki/Big_Ten_Conference}}

\examples{
data(big.eight)
}
\keyword{datasets}
