% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGPXtoNearbyOSM.R
\name{bringFeatureToOSM}
\alias{bringFeatureToOSM}
\title{Choose points on OSM feature which correspond to point on tracks.}
\usage{
bringFeatureToOSM(track, osm_features, threshold)
}
\arguments{
\item{track}{List of trajectory}

\item{osm_features}{List of OpenStreetMap features}

\item{threshold}{Distance threshold in meters.s}
}
\value{
Closest distance between \code{x} and \code{y}.
}
\description{
Calculates the closest distance between a feature (e.g. OSM lines) and a feature list (points of tracks).
Depending on the distance's threshold, either the closest point on the OSM linestring is chosen or the point of the track is taken if no OSM feature is close enough.
The function returns a track matched to the OSM features. The track is returned in crs EPSG:4326.
}
\examples{
track = readGPXFile("./inst/extdata/track_run.gpx")
bbox <- st_bbox(track)
osm_features = readOSMFiles(bbox, c("tertiary, secondary, sidewalk, footway"))
osm_track = bringFeatureToOSM(track, osm_features, 10)
}
