% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\docType{class}
\name{Bucket}
\alias{Bucket}
\title{Bucket}
\format{An R6 class object.}
\usage{
Bucket
}
\arguments{
\item{.output}{output format}
}
\value{



}
\description{
Bucket
}
\examples{
## new Bucket
b<-Bucket$new('ross-test',autoCreate=F)
## auto create when bucket not exist.
b<-Bucket$new('ross-test',autoCreate=T)

## create
b<-Bucket$new('ross-test')
# create bucket after init.
b$create()

## delete
b$delete
b$list()
b$rm()
b$rm(confirm=T)
b$rm('test-', confirm=T)

## acl
b$acl
b$acl <- "private"
b$acl <- "public-read"

## logging
b$logging
b$logging <- list(TargetBucket='ross-test', TargetPrefix='log-')
b$logging <- list(TargetBucket='ross-test')
b$logging <- list(TargetPrefix='log-')
b$logging <- list()
b$logging <- NULL

## website
b$website
b$website <- list(Suffix='index.html', Key='404.html')
b$website <- list(Suffix='index.html')
b$website <- list(Key='404.html')
b$website <- list()
b$website <- NULL

## referer
b$referer
b$referer <- list(AllowEmptyReferer=T, RefererList=c('*.igenecode.com', 'aliyun.com'))
b$referer <- list(AllowEmptyReferer=T)
b$referer <- list(RefererList=c('*.igenecode.com', 'aliyun.com'))
b$referer <- NULL
b$referer <- list()

 ## lifecycle

 b$lifecycle$add('upload_', Object.Days = 5)
 b$lifecycle$save()
 b$lifecycle
 # lifecycle auto save is off to speedup by default when using Bucket class.
 # Turn on
 b$lifecycle$autoSave <- T
 b$lifecycle$add('upload_', Object.Days = 7)
 b$lifecycle


 ## CORS

 b$cors$add('*', 'GET')
 b$cors$save()
 b$cors
 # cors auto save is off to speedup by default when using Bucket class.
 # Turn on
 b$cors$autoSave <- T
 b$cors$add('*', 'GET')
 b$cors

}
\seealso{
\code{\link{BucketLifecycle}}

\code{\link{BucketLifecycle}}
'
}
\keyword{datasets}
