% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cdnState}
\alias{cdnState}
\title{cdnState}
\usage{
cdnState(bucketname, domain)
}
\arguments{
\item{domain}{}
}
\description{
cdnState
}
\examples{
cdnState('igenecode-report', domain = 'report.igenecode.com')
cdnState('igenecode-report')
cdnState()
}
