% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipart-upload-api.R
\name{CompleteMultipartUpload}
\alias{CompleteMultipartUpload}
\title{CompleteMultipartUpload}
\usage{
CompleteMultipartUpload(bucketname, key, uploadId, Etags = NULL)
}
\arguments{
\item{uploadId}{Multipart upload event ID. InitiateMultipartUpload will return this in the response.}

\item{Etags}{PartNumber named character vector, which contains ETags of all parts.}
}
\description{
CompleteMultipartUpload
}
\examples{
CompleteMultipartUpload('ross-test', 'multi-test.txt', '300C834B2037432E871D62B848087139')
}
