% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{copyObjects}
\alias{copyObjects}
\title{copyObjects}
\usage{
copyObjects(src, dest, src_bucket = NULL, dest_bucket = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
copyObjects
}
\examples{
# Upload
copyObjects('tests/test_upload/multiplefiles/will_success/', 'success', dest_bucket = 'ross-test')
# Download
copyObjects('success', '/Volumes/RamDisk/', src_bucket = 'ross-test')
# Online Copy
copyObjects('success/1/1.txt', 'cp/1/1.txt', 'ross-test', 'ross-test')
}
