% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{createBucket}
\alias{createBucket}
\title{createBucket}
\usage{
createBucket(bucketname, Location = "oss-cn-beijing", acl = "private",
  StorageClass = "Standard")
}
\arguments{
\item{Location}{The region of bucket.}

\item{acl}{Authority control of bucket. choose from public-read-write, public-read, private.}

\item{StorageClass}{Storage type, Standard, IA or Archive.}
}
\description{
createBucket
}
\examples{
createBucket('ross-test')
}
