% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{DeleteBucketcors}
\alias{DeleteBucketcors}
\title{DeleteBucketcors}
\usage{
DeleteBucketcors(name)
}
\arguments{
\item{name}{}
}
\description{
DeleteBucketcors
}
\examples{
DeleteBucketcors('ross-test')
}
