% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{DeleteBucketLifecycle}
\alias{DeleteBucketLifecycle}
\title{DeleteBucketLifecycle}
\usage{
DeleteBucketLifecycle(name)
}
\arguments{
\item{name}{}
}
\description{
Delete all lifecycle rules for a bucket.
}
\examples{
DeleteBucketLifecycle('ross-test')
}
