% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{DeleteBucketLogging}
\alias{DeleteBucketLogging}
\title{DeleteBucketLogging}
\usage{
DeleteBucketLogging(name)
}
\arguments{
\item{name}{}
}
\description{
DeleteBucketLogging
}
\examples{
DeleteBucketLogging('ross-test')
}
