% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{DeleteBucketWebsite}
\alias{DeleteBucketWebsite}
\title{DeleteBucketWebsite}
\usage{
DeleteBucketWebsite(name)
}
\arguments{
\item{name}{}
}
\description{
DeleteBucketWebsite
}
\examples{
DeleteBucketWebsite('ross-test')
}
