% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{DeleteMultipleObjects}
\alias{DeleteMultipleObjects}
\title{DeleteMultipleObjects}
\usage{
DeleteMultipleObjects(bucketname, keys, quiet = TRUE)
}
\arguments{
\item{bucketname}{The bucketname}

\item{keys}{Object keys to be deleted. Should less than 1000 once.}

\item{quiet}{Return deleted keys or not. When TRUE, only return keys with errors.}
}
\description{
DeleteMultipleObjects
}
\examples{
r<-GetBucket('ross-test')
keys<-unlist(xpath2list(httr::content(r), '/ListBucketResult/Contents/Key'))
DeleteMultipleObjects('ross-test', keys, FALSE)
}
