% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{DeleteObject}
\alias{DeleteObject}
\title{DeleteObject}
\usage{
DeleteObject(bucketname, key)
}
\arguments{
\item{bucketname}{The bucket name}

\item{key}{The file path of object on oss.}
}
\description{
DeleteObject
}
\examples{
PutObject('ross-test', 'test.txt', 'test')
DeleteObject('ross-test', 'test.txt')
}
