% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{downloadMultipleObjects}
\alias{downloadMultipleObjects}
\title{downloadMultipleObjects}
\usage{
downloadMultipleObjects(bucketname, src, dest = ".", pattern = NULL,
  resume = TRUE, split = 5, method = "aria2", quiet = TRUE, ...,
  .progressbar = TRUE, .parallel = TRUE)
}
\arguments{
\item{src}{The objects to be downloaded. Ended with "/" means the whold "folder" will be download.}

\item{dest}{Local destination to save the files.}

\item{pattern}{Filter which files to be uploaded.}

\item{resume}{Auto resume from last failed download or not.}

\item{split}{How many download progress at the same time.}

\item{method}{Same argument in download.file. Supports aria2 if installed.}

\item{quiet}{Suppress status messages or not.}

\item{...}{Arguments pass to downloadObject.}

\item{.progressbar}{Show progress bar or not.}

\item{.parallel}{Parallel multiple upload or not. When False, split will be disable too.}
}
\value{
named status_codes indicates failed or success.
}
\description{
downloadMultipleObjects
}
\examples{
downloadMultipleObjects('ross-test', 'test/tmp')
r<-downloadMultipleObjects('ross-test', 'test/tmp/cache2/')
r<-downloadMultipleObjects('ross-test', 'test', '~/asdf')
r<-downloadMultipleObjects('ross-test', 'test', '~/asdf', resume=F, .parallel = F)
r<-downloadMultipleObjects('ross-test', 'test', '/Volumes/RamDisk/asdf', pattern="tmp", quiet = F, split=10, .progressbar = F)
}
