% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{downloadObject}
\alias{downloadObject}
\title{downloadObject}
\usage{
downloadObject(bucketname, src, dest = NULL, resume = TRUE, split = 5,
  method = "aria2", quiet = FALSE, extra = "", .md5 = T, rpc = NULL,
  ...)
}
\arguments{
\item{src}{The object path on bucket to be downloaded.}

\item{dest}{Local destination of file or folder.}

\item{resume}{Auto resume from failed download or not.}

\item{split}{How many parts to be split.}

\item{method}{Same argument in download.file. Supports aria2 if installed.}

\item{quiet}{Suppress status messages or not.}

\item{extra}{additional command-line arguments for the "wget", "curl" and "aria2" methods.}

\item{.md5}{Check md5 after download or not.}

\item{rpc}{aria2 rpc}

\item{...}{Arguments pass to download.file}
}
\description{
downloadObject
}
\examples{
downloadObject('ross-test', 'jingyi.pdf')
downloadObject('ross-test', 'jingyi.pdf', '~')
downloadObject('ross-test', 'jingyi.pdf', '~/jingyi2.pdf', resume = F, method = 'aria2', quiet = T)
downloadObject('ross-test', 'jingyi.pdf', '~/jingyi2.pdf', resume = T, method = 'wget', quiet = F)
}
