% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessData.R
\name{filterForClosePoints}
\alias{filterForClosePoints}
\title{Delete Outliers in Tracks}
\usage{
filterForClosePoints(track, threshold)
}
\arguments{
\item{track}{List of trajectory.}

\item{threshold}{Distance threshold. Unit based on crs of track of Points.}
}
\value{
Returns all entries of the \code{trackOfPoints} that are close enough.
}
\description{
This function selects points that are too far away from its next neighboring points.
}
\details{
The distance can be determined by the threshold.
Whenever the distance to one direct neighboring point is less than the threshold, the point will be kept.
It has the main purpose to identify outliers of GPS measurements when the receiver misslocated itself
}
\examples{
\dontrun{
track = readGPXFile("./inst/extdata/track_run.gpx")
filteredTrack = filterForClosePoints(track, 2)
}
}
