% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{GetBucket}
\alias{GetBucket}
\title{GetBucket}
\usage{
GetBucket(bucketname, prefix = NULL, marker = NULL, delimiter = NULL,
  max_keys = NULL, encoding_type = NULL)
}
\arguments{
\item{prefix}{The prefix of buckets to filter.}

\item{marker}{Which index to start with.}

\item{delimiter}{Single character grouping the object name.}

\item{max_keys}{Max number of buckets.}

\item{encoding_type}{For those control characters not supported in xml 1.0 should set this to `url`.}
}
\description{
Equals to ListObject
}
