% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{GetBucketLifecycle}
\alias{GetBucketLifecycle}
\title{GetBucketLifecycle}
\usage{
GetBucketLifecycle(name)
}
\arguments{
\item{name}{}
}
\description{
GetBucketLifecycle
}
\examples{
GetBucketLifecycle('ross-test')
}
