% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{GetBucketLocation}
\alias{GetBucketLocation}
\title{GetBucketLocation}
\usage{
GetBucketLocation(name)
}
\arguments{
\item{name}{}
}
\description{
GetBucketLocation
}
\examples{
GetBucketLocation('ross-test')
}
