% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{GetBucketLogging}
\alias{GetBucketLogging}
\title{GetBucketLogging}
\usage{
GetBucketLogging(name)
}
\arguments{
\item{name}{}
}
\description{
GetBucketLogging
}
\examples{
GetBucketLogging('ross-test')
}
