% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{GetBucketReferer}
\alias{GetBucketReferer}
\title{GetBucketReferer}
\usage{
GetBucketReferer(name)
}
\arguments{
\item{name}{}
}
\description{
GetBucketReferer
}
\examples{
GetBucketReferer('ross-test')
}
