% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{GetBucketWebsite}
\alias{GetBucketWebsite}
\title{GetBucketWebsite}
\usage{
GetBucketWebsite(name)
}
\arguments{
\item{name}{}
}
\description{
GetBucketWebsite
}
\examples{
GetBucketWebsite('ross-test')
}
