% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{GetObjectACL}
\alias{GetObjectACL}
\title{GetObjectACL}
\usage{
GetObjectACL(bucketname, key)
}
\arguments{
\item{bucketname}{The bucket name}

\item{key}{The file path of object on oss.}
}
\description{
GetObjectACL
}
\examples{
GetObjectACL('ross-test', 'test.txt')
}
