% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getObjectInfo}
\alias{getObjectInfo}
\title{getObjectInfo}
\usage{
getObjectInfo(bucketname, key, print = TRUE)
}
\arguments{
\item{key}{}
}
\description{
getObjectInfo
}
\examples{
PutObject('ross-test', 'test.txt')
getObjectInfo('ross-test', 'test.txt')
}
