% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{GetObjectMeta}
\alias{GetObjectMeta}
\title{GetObjectMeta}
\usage{
GetObjectMeta(bucketname, key)
}
\arguments{
\item{bucketname}{The bucket name}

\item{key}{The file path of object on oss.}
}
\description{
Only includes basic meta info like size, etag, etc. For complete meta use HeadObject.
}
\examples{
PutObject('ross-test', 'test.txt')
GetObjectMeta('ross-test', 'test.txt')
}
