% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{GetSymlink}
\alias{GetSymlink}
\title{GetSymlink}
\usage{
GetSymlink(bucketname, key)
}
\arguments{
\item{key}{}
}
\description{
GetSymlink
}
\examples{
GetSymlink('ross-test', 'test-linked.txt')
GetSymlink('ross-test', 'test.txt')
}
