% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{HeadObject}
\alias{HeadObject}
\title{HeadObject}
\usage{
HeadObject(bucketname, key, ETag = NULL, ETag.match = TRUE, since = NULL,
  modified.since = TRUE)
}
\arguments{
\item{bucketname}{The bucket name}

\item{key}{The file path of object on oss.}

\item{ETag}{ETag of object.}

\item{ETag.match}{Copy when ETag matches or copy when Etag not match.}

\item{since}{POSIXct time.}

\item{modified.since}{Copy when object was modified since a time, or copy when object was not modified since a time.}
}
\description{
HeadObject
}
\examples{
r <- PutObject('ross-test', 'test.txt')
HeadObject('ross-test', 'test.txt')
etag <- gsub('"', '', r$headers$etag)
HeadObject('ross-test', 'test.txt', ETag=etag, ETag.match=T)
HeadObject('ross-test', 'test.txt', ETag=etag, ETag.match=F)
}
