% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipart-upload-api.R
\name{InitiateMultipartUpload}
\alias{InitiateMultipartUpload}
\title{InitiateMultipartUpload}
\usage{
InitiateMultipartUpload(bucketname, key, encryption = NULL, ...)
}
\arguments{
\item{encryption}{server side encryption algorithm AES256}

\item{...}{Set header directly. Cache-Control, Content-Disposition, Content-Encoding, Expires, ...}
}
\description{
InitiateMultipartUpload
}
\examples{
InitiateMultipartUpload('ross-test', 'multi-test.txt')
}
