% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{isObjectExist}
\alias{isObjectExist}
\title{isObjectExist}
\usage{
isObjectExist(bucketname, key)
}
\arguments{
\item{key}{}
}
\description{
isObjectExist
}
\examples{
isObjectExist('ross-test', 'test.txt')
}
