% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{linkObject}
\alias{linkObject}
\title{linkObject}
\usage{
linkObject(bucketname, key, target, .meta = NULL)
}
\arguments{
\item{.meta}{}
}
\description{
linkObject
}
\examples{
PutObject('ross-test', 'test.txt')
linkObject('ross-test', 'linked-test.txt', '/ross-test/test.txt')
}
