% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{listBucket}
\alias{listBucket}
\title{listBucket}
\usage{
listBucket(bucketname, prefix = NULL, marker = NULL, delimiter = "/",
  max_keys = "1000", .all = TRUE, .output = "data.frame")
}
\arguments{
\item{prefix}{The prefix of buckets to filter.}

\item{marker}{Which index to start with.}

\item{delimiter}{Single character grouping the object name.}

\item{max_keys}{Max number of buckets.}

\item{.all}{No Paging}

\item{.output}{output format, choose from data.frame, list, oss, character. default is data.frame}
}
\description{
listBucket
}
\examples{
listBucket('ross-test')
listBucket('ross-test', 'upload')
listBucket('ross-test', 'upload', 'upload/file1', '/', '10')
}
