% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{listMultipartUploads}
\alias{listMultipartUploads}
\title{listMultipartUploads}
\usage{
listMultipartUploads(bucketname, prefix = NULL, delimiter = NULL,
  max = 1000, marker = NULL, id_marker = NULL, encoding_type = NULL)
}
\arguments{
\item{prefix}{The prefix of buckets to filter.}

\item{delimiter}{Single character grouping the object name.}

\item{max}{Max number of Multipart Upload events.}

\item{marker}{Which index to start with.}

\item{id_marker}{Should be use with marker parameter. Multipart Upload event which UploadId grater than id_marker will be return.}

\item{encoding_type}{For those control characters not supported in xml 1.0 should set this to `url`.}
}
\description{
listMultipartUploads
}
