% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipart-upload-api.R
\name{ListParts}
\alias{ListParts}
\title{ListParts}
\usage{
ListParts(bucketname, key, uploadId, max = 1000, marker = NULL)
}
\arguments{
\item{uploadId}{Multipart upload event ID. InitiateMultipartUpload will return this in the response.}

\item{max}{Max number of parts.}

\item{marker}{Which index to start with.}
}
\description{
ListParts
}
\examples{
r <- InitiateMultipartUpload('ross-test', 'multi-test.txt')
uploadId <- unlist(xpath2list(content(r), '//UploadId'))
UploadPart('ross-test', 'multi-test.txt', uploadId, 1, 'test1')
UploadPart('ross-test', 'multi-test.txt', uploadId, 2, 'test2')
ListParts('ross-test', 'multi-test.txt', uploadId)
ListParts('ross-test', 'multi-test.txt', uploadId, max=1)
ListParts('ross-test', 'multi-test.txt', uploadId, marker = 1)
AbortMultipartUpload('ross-test', 'multi-test.txt', uploadId)
}
