% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{loadObject}
\alias{loadObject}
\title{loadObject}
\usage{
loadObject(bucketname, key, ..., envir = parent.frame(), quiet = T)
}
\arguments{
\item{quiet}{}
}
\description{
loadObject
}
\examples{
loadObject('ross-test', 'test.RData')
ls()
}
