% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{locationState}
\alias{locationState}
\title{locationState}
\usage{
locationState(bucketname, location)
}
\arguments{
\item{location}{}
}
\description{
locationState
}
\examples{
locationState('ross-test')
locationState('ross-test', 'oss-cn-beijing')
}
