% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{metaMultipleObjects}
\alias{metaMultipleObjects}
\title{metaMultipleObjects}
\usage{
metaMultipleObjects(bucketname, prefix, meta, recursive = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
metaMultipleObjects
}
\examples{
PutObject('ross-test', 'test1.txt')
PutObject('ross-test', 'test2.txt')
metaMultipleObjects('ross-test', 'test1.txt', meta=list(a=1), .progressbar=F)
metaObject('ross-test', 'test1.txt')
metaMultipleObjects('ross-test', 'test', meta=list(b=2), recursive = T)
metaObject('ross-test', 'test1.txt')
metaObject('ross-test', 'test2.txt')
}
