% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{metaObject}
\alias{metaObject}
\title{metaObject}
\usage{
metaObject(bucketname, key, meta)
}
\arguments{
\item{meta}{}
}
\description{
metaObject
}
\examples{
PutObject('ross-test', 'test.txt')
metaObject('ross-test', 'test.txt', meta=list(test1='a', test2='b'))
metaObject('ross-test', 'test.txt')
metaObject('ross-test', 'test.txt', meta=list(test3='d'))
metaObject('ross-test', 'test.txt')
metaObject('ross-test', 'test.txt', meta=list(test3='c'))
metaObject('ross-test', 'test.txt')
metaObject('ross-test', 'test.txt', meta=list(test3=NULL))
metaObject('ross-test', 'test.txt')
metaObject('ross-test', 'test.txt', meta=NULL)
metaObject('ross-test', 'test.txt')
}
