% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.acl}
\alias{oss.acl}
\title{oss.acl}
\usage{
oss.acl(x, ...)
}
\arguments{
\item{...}{}
}
\description{
oss.acl
}
\examples{
oss.acl('oss://ross-test')
oss.acl('oss://ross-test', acl='public-read')

oss.acl('oss://ross-test/test.txt')
oss.acl('oss://ross-test/test.txt', acl='private')
}
