% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.exists}
\alias{oss.exists}
\title{oss.exists}
\usage{
oss.exists(x)
}
\arguments{
\item{x}{}
}
\description{
oss.exists
}
\examples{
oss.exists('oss://ross-test')
oss.exists('oss://ross-test/test.txt')

o <- Object$new('ross-test', 'test.txt')
oss.exists(o)

b <- Bucket$new('ross-test')
oss.exist(b)
}
