% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.load}
\alias{oss.load}
\title{oss.load}
\usage{
oss.load(x, ...)
}
\arguments{
\item{...}{}
}
\description{
oss.load
}
\examples{
oss.load('oss://ross-test/test.RData')
ls()
e <- new.env()
oss.load('oss://ross-test/test.RData', envir=e)
ls(e)
}
