% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.ls}
\alias{oss.ls}
\title{oss.ls}
\usage{
oss.ls(x, ...)
}
\arguments{
\item{...}{}
}
\description{
oss.ls
}
\examples{
oss.ls()
bl <- BucketList$new()
oss.ls(bl)

oss.ls(oss('oss://ross-test'))
oss.ls('oss://ross-test')
b <- Bucket$new('ross-test')
oss.ls(b)
}
