% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.mb}
\alias{oss.mb}
\title{oss.mb}
\usage{
oss.mb(x, ...)
}
\arguments{
\item{...}{}
}
\description{
oss.mb
}
\examples{
oss.mb(oss('oss://ross-test'))
oss.mb('oss://ross-test')
b <- Bucket$new('ross-test')
oss.mb(b)
}
