% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.meta}
\alias{oss.meta}
\title{oss.meta}
\usage{
oss.meta(x, ...)
}
\arguments{
\item{...}{}
}
\description{
oss.meta
}
\examples{
oss.meta('oss://ross-test/test.txt')
oss.meta('oss://ross-test/test.txt', meta=list(b=2))
oss.meta('oss://ross-test/test.txt', meta=list(a=1))
oss.meta('oss://ross-test/test.txt', meta=list(a=NULL))
oss.meta('oss://ross-test', meta=NULL)
oss.meta('oss://ross-test/test/', recursive=T, meta=list(a=1))
}
