% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.read}
\alias{oss.read}
\title{oss.read}
\usage{
oss.read(x, ...)
}
\arguments{
\item{...}{}
}
\description{
oss.read
}
\examples{
oss.read('oss://ross-test/test.txt')
oss.read('oss://ross-test/test.txt', Range='0-100')
}
