% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.readRDS}
\alias{oss.readRDS}
\title{oss.readRDS}
\usage{
oss.readRDS(x, ...)
}
\arguments{
\item{...}{}
}
\description{
oss.readRDS
}
\examples{
oss.readRDS('oss://ross-test/test.rds')
e <- new.env()
oss.readRDS('oss://ross-test/test.rds', envir=e)
}
