% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.restore}
\alias{oss.restore}
\title{oss.restore}
\usage{
oss.restore(x)
}
\arguments{
\item{x}{}
}
\description{
oss.restore
}
\examples{
oss.mb('oss://ross-test-arch', StorageClass='Archive')
oss.write('oss://ross-test-arch/test.txt', 'test')
oss.restore('oss://ross-test-arch/test.txt')

o <- Object$new('ross-test-arch', 'test.txt')
oss.restore(o)
}
