% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss.usage}
\alias{oss.usage}
\title{oss.usage}
\usage{
oss.usage(x, ...)
}
\arguments{
\item{...}{}
}
\description{
oss.usage
}
\examples{
oss.usage('oss://ross-test/')
oss.usage('oss://ross-test/', unit='KB')
oss.usage('oss://ross-test/test/')
oss.usage('oss://ross-test/test/test.txt')
}
