% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGPXtoNearbyOSM.R
\name{points2line_trajectory}
\alias{points2line_trajectory}
\title{Point to Line Conversion in List}
\usage{
points2line_trajectory(p)
}
\arguments{
\item{p}{points}
}
\value{
The points of \code{p} merged to linestrings starting with \code{p[i-1]} and ending with \code{p[i]}.
}
\description{
Convert a list of observations of Points to a list of observations with lines between points.
The conversion adds the parameter speed.
}
\examples{
points2line_trajectory(c(st_point(c(0,1)), st_point(c(1,2)), st_point(c(4,1)), st_point(c(4,2))))
}
