% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{PutBucket}
\alias{PutBucket}
\title{PutBucket}
\usage{
PutBucket(bucketname, Location = "oss-cn-beijing", acl = "private",
  StorageClass = "Standard")
}
\arguments{
\item{Location}{The region of bucket.}

\item{acl}{Authority control of bucket. choose from public-read-write, public-read, private.}

\item{StorageClass}{Storage type, Standard or IA.}

\item{name}{The bucket name}
}
\description{
Create new Bucket. Modify authority control.
}
\examples{
PutBucket('ross-test', 'public-read-write')
}
