% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{PutBucketcors}
\alias{PutBucketcors}
\title{PutBucketcors}
\usage{
PutBucketcors(name, rules, body = NULL)
}
\arguments{
\item{name}{The bucket name}

\item{rules}{list Rules created by `.build.xml_body.PutBucketLifecycle.Rules`. Will be ignore when body parameter is specified.}

\item{body}{charactor XML content of rules.}
}
\description{
PutBucketcors
}
\examples{
rules <- list()
rules[[1]] <- .build.xml_body.PutBucketcors.Rules(AllowedOrigin="*", AllowedMethod='GET')
rules[[2]] <- .build.xml_body.PutBucketcors.Rules(AllowedOrigin="igenecode.com", AllowedMethod=c('GET', 'PUT'))
rules[[3]] <- .build.xml_body.PutBucketcors.Rules(AllowedOrigin="www.igenecode.com", AllowedMethod='GET',
 AllowedHeader='Authorization', ExposeHeader=c('x-oss-meta1', 'x-oss-meta2'), MaxAgeSeconds=100)
PutBucketcors('ross-test', rules)
}
