% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{PutBucketLifecycle}
\alias{PutBucketLifecycle}
\title{PutBucketLifecycle}
\usage{
PutBucketLifecycle(name, rules, body = NULL)
}
\arguments{
\item{name}{The bucket name}

\item{Prefix}{Object prefix applying the rule.}

\item{RuleID}{Uid of rule}

\item{Status}{Enable or Disable the rule.}

\item{Object.CreatedBeforeDate}{Expires date for object.}

\item{Object.Days}{Expris days for object.}

\item{Multpart.CreatedBeforeDate}{Expris date for multipart upload.}

\item{Multpart.Days}{Expris day for multipart upload.}
}
\description{
Config bucket object life cycle rules.
}
\examples{

rules <- list()

rules[[1]] <- .build.xml_body.PutBucketLifecycle.Rules(Prefix = 'upload_', Object.CreatedBeforeDate = Sys.Date()+7)
rules[[2]] <- .build.xml_body.PutBucketLifecycle.Rules(Prefix = 'upload_', Multpart.Days = 5)
rules[[3]] <- .build.xml_body.PutBucketLifecycle.Rules(Prefix = 'upload_', Object.Days = 30, Multpart.Days = 5)
rules[[4]] <- .build.xml_body.PutBucketLifecycle.Rules(Prefix = 'upload_', Object.Days = 30)

PutBucketLifecycle('ross-test', rules)
}
