% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{PutBucketLogging}
\alias{PutBucketLogging}
\title{PutBucketLogging}
\usage{
PutBucketLogging(name, prefix, target = name, on = TRUE)
}
\arguments{
\item{name}{The bucket name}

\item{prefix}{The prefix of log object.}

\item{target}{The target bucket saving the log object.}
}
\description{
Turn on/off bucket logging.
}
\examples{
r <- PutBucketLogging('ross-test', 'log-')
}
