% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{PutBucketReferer}
\alias{PutBucketReferer}
\title{PutBucketReferer}
\usage{
PutBucketReferer(name, AllowEmptyReferer = TRUE, RefererList = c())
}
\arguments{
\item{name}{The bucket name}

\item{AllowEmptyReferer}{Allow empty referer or not. boolean}

\item{RefererList}{Referer white list.}
}
\description{
Configure referer white list.
}
\examples{
PutBucketReferer('ross-test', AllowEmptyReferer=FALSE, RefererList='http://*.aliyun.com')
}
