% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{PutBucketWebsite}
\alias{PutBucketWebsite}
\title{PutBucketWebsite}
\usage{
PutBucketWebsite(name, suffix = "index.html", key = "404.html")
}
\arguments{
\item{name}{The bucket name}

\item{suffix}{Index file name.}

\item{key}{404 error file name.}
}
\description{
PutBucketWebsite
}
\examples{
PutBucketWebsite('ross-test')
PutBucketWebsite('ross-test', key='error.html', suffix='index.html')
}
