% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{PutObject}
\alias{PutObject}
\title{PutObject}
\usage{
PutObject(bucketname, key, body = "", encryption = NULL, acl = "private",
  ..., .md5 = TRUE, .meta = NULL, .overwrite = TRUE)
}
\arguments{
\item{bucketname}{The bucket name}

\item{key}{The file path of object on oss.}

\item{body}{The content of object, character or upload_file(file_path).}

\item{encryption}{server side encryption algorithm AES256}

\item{acl}{Access control: private, public-read, public-read-write}

\item{...}{Set header directly. Cache-Control, Content-Disposition, Content-Encoding, Expires, ...}

\item{.md5}{boolean Calculate body md5sum automatically or not.}

\item{.meta}{Other meta info set to the object, < 8k.}

\item{.overwrite}{overwrite or not. default TRUE}
}
\description{
PutObject
}
\examples{
PutObject('ross-test', 'test.txt', 'test')
PutObject('ross-test', 'test.txt', 'test', encryption = 'AES256')
PutObject('ross-test', 'test.txt', 'test', acl = 'public-read')
PutObject('ross-test', 'test.txt', 'test', .md5 = F)
PutObject('ross-test', 'test.txt', 'test', .meta = list(location='beijing', owner='igenecode.com'))
PutObject('ross-test', 'test.txt', 'test', .overwrite = F)
}
