% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{PutObjectACL}
\alias{PutObjectACL}
\title{PutObjectACL}
\usage{
PutObjectACL(bucketname, key, acl = "private")
}
\arguments{
\item{bucketname}{The bucket name}

\item{key}{The file path of object on oss.}

\item{acl}{Access control: private, public-read, public-read-write}
}
\description{
PutObjectACL
}
\examples{
PutObjectACL('ross-test', 'test.txt', acl = 'public-read')
}
