% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-api.R
\name{PutSymlink}
\alias{PutSymlink}
\title{PutSymlink}
\usage{
PutSymlink(bucketname, key, target, .meta = NULL)
}
\arguments{
\item{.meta}{}
}
\description{
PutSymlink
}
\examples{
PutObject('ross-test', 'test.txt')
PutSymlink('ross-test', 'test-linked.txt', 'test.txt')
}
