% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessData.R
\name{readGPXFile}
\alias{readGPXFile}
\title{Returns content of a GPX file}
\usage{
readGPXFile(fp)
}
\arguments{
\item{fp}{Path to GPX file.}
}
\value{
content of GPX file.
}
\description{
source: https://geocompr.github.io/geocompkg/articles/gps-tracks.html?tag=makemoney0821-20
}
\examples{
track = readGPXFile("./inst/extdata/track_run.gpx")
}
