% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readHomer.R
\name{readHomer}
\alias{readHomer}
\title{A function to read a HOMER-formatted peak file into a GRanges object}
\usage{
readHomer(file, header.regex = "^#PeakID", meta.cols = NULL,
  keep.all.metadata = FALSE, remove.unusual = FALSE)
}
\arguments{
\item{file}{an absolute or relative path to a HOMER-formatted peak file}

\item{header.regex}{an optional regular expression to identify the header line (Default:"^#PeakID").}

\item{meta.cols}{\code{named list} that maps column numbers to meta data columns. e.g. list(name=5, score=10), which means 5th column will be named "name", and 10th column will be named "score" and their contents will be a part of the returned \code{GRanges object}.}

\item{keep.all.metadata}{\code{logical} determining if the extra columns ( the ones that are not designated by chr,start,end,strand and meta.cols arguments ) should be kept or not. (Default:\code{FALSE})}

\item{remove.unusual}{if \code{TRUE}(default) remove the chromosomes with unsual names, such as chrX_random. (Default:\code{FALSE})}
}
\value{
\code{\link{Granges}} object.
}
\description{
A function to read a HOMER formatted peak/position file into a GRanges object. HOMER is a popular DNA motif analysis and general NGS program suite.
Peak/position files have a commented header and are formatted as:
PeakID, chr, start, end, strand ...
}
\examples{
# peaks <- readHomer("peaks.txt", keep.all.metadata=TRUE)
# head(peaks)
}
\seealso{
\code{\link{readGeneric}}
}

