% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessData.R
\name{readOSMFiles}
\alias{readOSMFiles}
\title{Returns all OSM features within a bounding box}
\usage{
readOSMFiles(bbox)
}
\arguments{
\item{bbox}{bounding box. Must be ordered like this: left, bottom, right top}
}
\value{
list of requested features cropped by the bounding box.
}
\description{
Returns all OSM features within a bounding box
}
\examples{
track = readGPXFile("./inst/extdata/track_run.gpx")
bbox <- st_bbox(track)
osm_features = readOSMFiles(bbox)
}
