% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{readRDSObject}
\alias{readRDSObject}
\title{readRDSObject}
\usage{
readRDSObject(bucketname, key, ..., refhook = NULL, quiet = T)
}
\arguments{
\item{quiet}{}
}
\description{
readRDSObject
}
\examples{
readRDSObject('ross-test', 'test.rds')
}
