% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{removeObjects}
\alias{removeObjects}
\title{removeObjects}
\usage{
removeObjects(bucketname, prefix = NULL, confirm = FALSE, quiet = TRUE,
  step = 1000)
}
\arguments{
\item{prefix}{The prefix of objects to be deleted. NULL means all objects in the bucket. Prefix ended with "/" means it's a folder.}

\item{confirm}{Auto confirm deletion or not.}

\item{step}{How many to delete at a time.}
}
\description{
removeObjects
}
\examples{
removeObjects('ross-test')
removeObjects('ross-test', 'upload/')
removeObjects('ross-test', 'upload/', confirm=TRUE)
}
