% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{restoreObject}
\alias{restoreObject}
\title{restoreObject}
\usage{
restoreObject(bucketname, key)
}
\arguments{
\item{key}{}
}
\description{
restoreObject
}
\examples{
createBucket('ross-test-archive', StorageClass = 'Archive')
PutObject('ross-test-archive', 'test.txt')
restoreObject('ross-test-archive', 'test.txt')
}
