% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ross.R
\docType{package}
\name{ross}
\alias{ross}
\alias{ross-package}
\title{ross: ross is an aliyun OSS API Wrapper for R.}
\description{
The ross package provides the basic OSS API. includes:
Bucket, Object, STS
}
\section{Foo functions}{

The foo functions ...
}

