% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGPXtoNearbyOSM.R
\name{saveGPXFile}
\alias{saveGPXFile}
\title{A trajectory gets analyzed with respect to movement statistics.}
\usage{
saveGPXFile(track, filePath)
}
\arguments{
\item{track}{List of trajectory}

\item{filePath}{String with path where GPX file shall be stored}
}
\value{
Evaluation plots of track
}
\description{
The main purpose is an evaluation of trajectories tracked during sports.
}
\examples{
\dontrun{
 saveGPXFile(track, "./inst/extdata/track_run_special.gpx")
}

}
