% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{saveObject}
\alias{saveObject}
\title{saveObject}
\usage{
saveObject(bucketname, key, ..., envir = parent.frame(), opts = NULL)
}
\arguments{
\item{opts}{arguments pass to uploadObject.}
}
\description{
saveObject
}
\examples{
a <- 1:4
b <- 5:6
saveObject('ross-test', 'test.RData', a, b)
saveObject('ross-test', 'test.RData', a, b, opts=list(split = 10))
}
