% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{saveRDSObject}
\alias{saveRDSObject}
\title{saveRDSObject}
\usage{
saveRDSObject(bucketname, key, object, ..., opts = NULL)
}
\arguments{
\item{opts}{}
}
\description{
saveRDSObject
}
\examples{
saveRDSObject('ross-test', 'test.rds', 1:5)
}
