\name{top.twenty}
\alias{top.twenty}
\docType{data}
\title{MTV Top 20 Greatest Music Videos Ever Made}
\description{
Compiled by MTV in 1999, this dataset includes the song title of the top 20 greatest music videos ever made.
}
\usage{data(top.twenty)}
\format{
  A data frame with 20 observations on the following variable.
  \describe{
    \item{\code{Video}}{a factor with levels \code{Addicted to Love} \code{Beat It} \code{California Love} \code{Closer} \code{Every Breath You Take} \code{Express Yourself} \code{Freedom} \code{Hungry Like the Wolf} \code{Jeremy} \code{Losing My Religion} \code{Put Your Hands Where My Eyes Can See} \code{Sabotage} \code{Sledgehammer} \code{Smells Like Teen Spirit} \code{Sweet Child O Mine} \code{Take On Me} \code{The Rain (Supa Dupa Fly)} \code{Thriller} \code{Vogue} \code{Walk This Way}}
  }
}

\source{\url{http://www.rockonthenet.com/archive/1999/mtv100.htm}}

\examples{
data(top.twenty)

}
\keyword{datasets}
