% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{uploadMultipleObjects}
\alias{uploadMultipleObjects}
\title{uploadMultipleObjects}
\usage{
uploadMultipleObjects(bucketname, src, prefix = "/", pattern = NULL,
  resume = TRUE, split = 5, .progressbar = TRUE, ..., .parallel = TRUE)
}
\arguments{
\item{src}{Path to local file or dir to be uploaded.}

\item{prefix}{Destination on Bucket. Ended with "/" means folder, files will place under "prefix/src/*". otherwise will be "prefix/*".}

\item{pattern}{Filter which files to be uploaded.}

\item{resume}{Auto resume from last failed upload or not.}

\item{split}{How many upload progress at the same time.}

\item{.progressbar}{Show progress bar or not.}

\item{...}{Arguments pass to uploadObject.}

\item{.parallel}{Parallel multiple upload or not. When False, split will be disable too.}
}
\value{
named status_codes indicates failed or success.
}
\description{
uploadMultipleObjects
}
\examples{
uploadMultipleObjects('ross-test', 'R',  .parallel = F)
uploadMultipleObjects('ross-test', 'R', 'test', .parallel = F)
uploadMultipleObjects('ross-test', 'R', 'test/', .parallel = T)
}
