% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipart-upload-api.R
\name{UploadPart}
\alias{UploadPart}
\title{UploadPart}
\usage{
UploadPart(bucketname, key, uploadId, partNumber, body = NULL, ...,
  .md5 = TRUE)
}
\arguments{
\item{uploadId}{Multipart upload event ID. InitiateMultipartUpload will return this in the response.}

\item{partNumber}{Part number indicating which part uploading. Coresponding to relative position in the file. Data will be overwrite when the part already exists.}

\item{body}{The content of object, character or upload_file(file_path).}

\item{...}{Set header directly. Cache-Control, Content-Disposition, Content-Encoding, Expires, ...}

\item{.md5}{boolean Calculate body md5sum automatically or not.}
}
\description{
UploadPart
}
\examples{
InitiateMultipartUpload('ross-test', 'multi-test.txt')
UploadPart('ross-test', 'multi-test.txt', 'F7376AAB033344C1922A588D15CE56A2', 1, 'test')
}
