% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipart-upload-api.R
\name{UploadPartCopy}
\alias{UploadPartCopy}
\title{UploadPartCopy}
\usage{
UploadPartCopy(source, bucketname, key, uploadId, partNumber, Range = NULL,
  ETag = NULL, ETag.match = TRUE, since = NULL, modified.since = TRUE)
}
\arguments{
\item{source}{The source object. /bucketname/objectname}

\item{bucketname}{Destinate bucket}

\item{key}{Destinate path on bucket}

\item{uploadId}{Multipart upload event ID. InitiateMultipartUpload will return this in the response.}

\item{partNumber}{Part number indicating which part uploading. Coresponding to relative position in the file. Data will be overwrite when the part already exists.}

\item{Range}{The object content range, '0-99' means the first 100 bytes.}

\item{ETag}{ETag of object.}

\item{ETag.match}{Copy when ETag matches or copy when Etag not match.}

\item{since}{POSIXct time.}

\item{modified.since}{Copy when object was modified since a time, or copy when object was not modified since a time.}
}
\description{
UploadPartCopy
}
\examples{
PutObject('ross-test', 'test.txt', 'test')
r <- InitiateMultipartUpload('ross-test', 'multi-test.txt')
uploadId <- unlist(xpath2list(httr::content(r), '//UploadId'))
UploadPartCopy('/ross-test/test.txt', 'ross-test', 'multi-test.txt', uploadId, 2, Range = '0-1')
}
