% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{usageBucket}
\alias{usageBucket}
\title{usageBucket}
\usage{
usageBucket(bucketname, prefix = NULL, unit = "MB")
}
\arguments{
\item{prefix}{The prefix of objects. NULL means the whole bucket.}

\item{unit}{return in which unit, B, KB, MB, GB, TB}
}
\description{
Check disk usage of buckets or "folder".
}
\examples{
usageBucket('ross-test')
usageBucket('ross-test', 'upload/')
}
