% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{xpath2list}
\alias{xpath2list}
\title{xpath2list}
\usage{
xpath2list(doc, xpath = NULL, smart = TRUE)
}
\arguments{
\item{doc}{}
}
\description{
Turn xml nodes into list, xpath is supported
}
\examples{
r <- GetService()
xpath2list(r)
xpath2list(r, '/ListAllMyBucketsResult/Owner')
xpath2list(r, '/ListAllMyBucketsResult/Buckets/Bucket')
}
