% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{abortMultipartUpload}
\alias{abortMultipartUpload}
\title{abortMultipartUpload}
\usage{
abortMultipartUpload(bucketname, prefix = NULL)
}
\arguments{
\item{prefix}{The prefix of objects on bucket.}
}
\description{
Abort failed multipart uploads.
}
\examples{
abortMultipartUpload('ross-test', 'some-failed-multipart.gz')
abortMultipartUpload('ross-test')
}
