% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-api.R
\name{GetService}
\alias{GetService}
\alias{listBucket}
\title{GetService}
\usage{
GetService(prefix = NULL, marker = NULL, max_keys = 100)

listBucket(...)
}
\arguments{
\item{prefix}{The prefix of buckets to filter.}

\item{marker}{Which index to start with.}

\item{max_keys}{Max number of buckets.}
}
\description{
Equals to ListObject

ListBucket
}
\section{Functions}{
\itemize{
\item \code{listBucket}: 
}}

