% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket-api.R
\name{PutBucketLifecycle}
\alias{PutBucketLifecycle}
\title{PutBucketLifecycle}
\usage{
PutBucketLifecycle(name, rules, body = NULL)
}
\arguments{
\item{name}{The bucket name}

\item{rules}{list Rules created by `.build.xml_body.PutBucketLifecycle.Rules`. Will be ignore when body parameter is specified.}

\item{body}{charactor XML content of rules.}
}
\description{
Config bucket object life cycle rules.
}
\examples{

rules <- list()

rules[[1]] <- .build.xml_body.PutBucketLifecycle.Rules(Prefix = 'upload_', Object.CreatedBeforeDate = Sys.Date()+7)
rules[[2]] <- .build.xml_body.PutBucketLifecycle.Rules(Prefix = 'upload_', Multpart.Days = 5)
rules[[3]] <- .build.xml_body.PutBucketLifecycle.Rules(Prefix = 'upload_', Object.Days = 30, Multpart.Days = 5)
rules[[4]] <- .build.xml_body.PutBucketLifecycle.Rules(Prefix = 'upload_', Object.Days = 30)

PutBucketLifecycle('ross-test', rules)
}
